+++ rororo/openapi/constants.py
@@ -1,6 +1,9 @@
 #: Key to store valid OpenAPI context within the ``web.Request`` instance
 OPENAPI_CONTEXT_REQUEST_KEY = "rororo_openapi_context"
 
+#: Key to store necessity of validating responses against OpenAPI schema
+OPENAPI_IS_VALIDATE_RESPONSE_APP_KEY = "rororo_openapi_is_validate_response"
+
 #: Key to store OpenAPI schema within the ``web.Application`` instance
 OPENAPI_SCHEMA_APP_KEY = "rororo_openapi_schema"